requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
filelock==3.10.7
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.2
packaging==23.0
Pillow==9.4.0
pkg_resources==0.0.0
protobuf==4.22.1
pydantic==1.10.7
Pygments==2.14.0
PyPDF2==3.0.1
python-telegram-bot==20.2
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sniffio==1.3.0
soupsieve==2.4
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.2
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.3
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wikipedia==1.4.0
yarl==1.8.2
